Refactor dummy project fixtures to conftest.py
